refactor Primitive
